MongoDB 3.6 compatible version of pattern
