+++ api/materialized_views/consts.js
@@ -1,0 +1,2 @@
+const minDate       = new Date('1970-01-01T00:00:00Z');
+exports.minDate = minDate;
