+++ app_helper.js
@@ -3,7 +3,7 @@ const cron          = require('node-cron');
 const _             = require('lodash');
 const winston       = require('winston');
 const options       = require('./config/mongoose_options').mongooseOptions;
-const minDate       = new Date('1970-01-01T00:00:00Z');
+const materialized_view__top_search_terms       = require('./api/materialized_views/reports/top_search_terms');
 
 // Logging middleware
 winston.loggers.add('default', {
@@ -87,86 +87,15 @@ async function loadMongoose() {
   await loadModels(dbConnection, options, defaultLog);
 }
 
-async function update__read_only__reports__top_search_terms(defaultLog, afterTimestamp) {
-  let queryAggregates = [
-    {$project:{
-      "meta":"$meta",
-      "action":"$action",
-      "timestamp":"$timestamp"
-    }},
-    {$match:{
-      $and:[
-        {$or:[
-          {"action":{$eq:"search"}},
-          {"action":{$eq:"Search"}}]},
-        {"timestamp":{$gt: new Date(afterTimestamp) }}
-      ]
-    }},
-    {$project:{
-      "_id":"$_id",
-      "___group":{"meta":"$meta"},
-      "timestamp":"$timestamp"
-    }},
-    {$group:{
-      "_id":"$___group",
-      "count":{$sum:1},
-      "latest":{$max:"$timestamp"}
-    }},
-    {$project:{
-      "_id": {$ifNull: ["$_id.meta", "\\nnull\\n"]},
-      "count": true,
-      "latest": true
-    }}
-  ];
-  if (minDate != afterTimestamp) {
-    defaultLog.debug('checking if need to update read_only__reports__top_search_terms');
-    let latestSinceLastRun = await mongoose.model('Audit').aggregate(queryAggregates);
-    latestSinceLastRun.map((recentlyUpdated) => {
-      const collection = mongoose.connection.db.collection('read_only__reports__top_search_terms');
-      collection.updateOne({
-        "_id":recentlyUpdated['_id'],
-        "latest":{"$ne": recentlyUpdated['latest']}
-      },
-      { $inc: {
-        "count": recentlyUpdated['count']}
-      });
-      collection.updateOne({
-        "_id":recentlyUpdated['_id'],
-        "latest":{"$ne": recentlyUpdated['latest']}
-      },
-      { $set: {
-        "latest": recentlyUpdated['latest']}
-      });
-      defaultLog.debug('updated "' + recentlyUpdated['_id'] + '" to \'' + recentlyUpdated['latest'] + '\' and incremented count by ' + recentlyUpdated['count'] +']');
-    });
-  } else {
-    defaultLog.debug('initializing read_only__reports__top_search_terms');
-    queryAggregates.push({ $merge: { into: "read_only__reports__top_search_terms", whenMatched: "replace" } });
-    await mongoose.model('Audit').aggregate(queryAggregates);
-    const collection = mongoose.connection.db.collection('read_only__reports__top_search_terms');
-    collection.createIndex({_id: 1});
-    collection.createIndex({count: 1});
-    collection.createIndex({latest: 1});
-  }
-}
-
-async function get_last_update__top_search_terms(defaultLog) {
-  const collection = mongoose.connection.db.collection('read_only__reports__top_search_terms');
-  if (!collection) return minDate;
-  let result = await collection.find({}, {"_id":0, "latest":1}).sort({"latest":-1}).limit(1).toArray();
-  if (0 == result.length) return minDate;
-  defaultLog.debug('last update to search terms was: ' + new Date(result[0].latest));
-  return new Date(result[0].latest);
-}
-
 async function startCron(defaultLog) {
   // standard crom pattern
   // seconds[0-59] minutes[0-59] hours[0-23] day_of_month[1-31] months[0-11] day_of_week[0-6]
   // cron.schedule('* 3 * * * *', async function() {
   cron.schedule('10 * * * * *', async function() {
-    let afterTimestamp = await get_last_update__top_search_terms(defaultLog);
-    await update__read_only__reports__top_search_terms(defaultLog, afterTimestamp);
+    let afterTimestamp = await materialized_view__top_search_terms.get_last(defaultLog);
+    await materialized_view__top_search_terms.update(defaultLog, afterTimestamp);
   });
+  // TODO: add more materialized view cron tasks here
 }
 
 exports.loadMongoose = loadMongoose;
